+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [4.3.3] - UNRELEASED
+## [4.3.3] - 03 Feb, 2025
 
 ### Fixed
 
 - Fix the overdue actions check in the Scheduled Actions list screen (Issue #1155).
+- Fix the Send Email step to correctly parse the recipient field when using custom email addresses (Issue #1153).
 - Update translations (Issue #1156).